--- Fixed “double-free” error in `ClientDebuglayer` destructor
Added `Timer` class
Added a file dialogue system

Added Changelog file

Moved ImGui Debug window to dedicated `ClientDebugLayer`
Changed `OrthographicCameraController` class’ `OnUpdate()` function to pan with RMB instead of WASD
`OrhtographicCameraController` now requires `Framebuffer`.
`OrhtographicCameraController` uses framebuffer for size and aspect ratio for projection and mouse panning
Changed controls of `ClientNetLayer` to ImGui buttons instead of keyboard controls
Removed all `termios` code from Client code as all controls are now handled through the engine
Changed text regarding average frame time and rate
Changed logging format

+++ Engine/include/Engine/Event.h
@@ -192,8 +192,8 @@ namespace cee
 			{
 			}
 				
-			uint32_t GetX() const { return m_MouseX; }
-			uint32_t GetY() const { return m_MouseY; }
+			float GetX() const { return m_MouseX; }
+			float GetY() const { return m_MouseY; }
 			
 			std::string ToString() const override
 			{
@@ -217,8 +217,8 @@ namespace cee
 			{
 			}
 				
-			uint32_t GetXOffset() const { return m_XOffset; }
-			uint32_t GetYOffset() const { return m_YOffset; }
+			float GetXOffset() const { return m_XOffset; }
+			float GetYOffset() const { return m_YOffset; }
 			
 			std::string ToString() const override
 			{
